Refactor away from visitor approach
