+++ packages/gatsby/src/utils/babel/__tests__/fixtures/remove-apis/destructured-export/input.mjs
@@ -10,6 +10,7 @@ export const {
   config,
   config2,
 } = () => {
+  const { config } = 'hello'; // Ensure matching nested properties stay untouched
   return {
     defer: true,
   }
