+++ packages/gatsby/src/utils/babel/__tests__/fixtures/remove-apis/destructured-export/output.mjs
@@ -1,6 +1,10 @@
 export const {
   config2
 } = () => {
+  const {
+    config
+  } = 'hello'; // Ensure matching nested properties stay untouched
+
   return {
     defer: true
   };
