+++ packages/gatsby/src/utils/babel/babel-plugin-remove-api.ts
@@ -1,7 +1,7 @@
 import { declare } from "@babel/helper-plugin-utils"
 import * as t from "@babel/types"
-import { PluginObj, ConfigAPI } from "@babel/core"
-import { RemoveNamedExportPropertiesVisitor } from "./babel-module-export-visitors"
+import { PluginObj, ConfigAPI, NodePath } from "@babel/core"
+import { removeExportProperties } from "./babel-module-exports"
 
 export default declare(function removeApiCalls(
   api: ConfigAPI,
@@ -105,6 +105,8 @@ export default declare(function removeApiCalls(
             | t.ExportNamespaceSpecifier
             | t.ExportSpecifier
           > = []
+
+          // Remove `export { foo }` shaped exports
           path.node.specifiers.forEach(specifier => {
             if (
               t.isExportSpecifier(specifier) &&
@@ -121,9 +123,13 @@ export default declare(function removeApiCalls(
         }
 
         let apiToCheck
+
+        // Remove `export function foo() {}` shaped exports
         if (t.isFunctionDeclaration(declaration) && declaration.id) {
           apiToCheck = declaration.id.name
         }
+
+        // Remove `export const foo = () => {}` shaped exports
         if (
           t.isVariableDeclaration(declaration) &&
           t.isIdentifier(declaration.declarations[0].id)
@@ -136,13 +142,20 @@ export default declare(function removeApiCalls(
           path.remove()
         }
 
-        path.traverse(RemoveNamedExportPropertiesVisitor, {
-          path,
-          propertiesToRemove: apisToRemove,
-        })
+        // Remove `export const { foo } = () => {}` shaped exports
+        if (t.isVariableDeclaration(declaration)) {
+          for (let i = 0; i < declaration?.declarations.length; i++) {
+            if (declaration?.declarations[i].id.type === `ObjectPattern`) {
+              const objectPath = path.get(
+                `declaration.declarations.${i}.id`
+              ) as NodePath<t.ObjectPattern>
+              removeExportProperties(path, objectPath, apisToRemove)
+            }
+          }
+        }
       },
 
-      // remove exports
+      // Remove `module.exports = { foo }` and `exports.foo = {}` shaped exports
       ExpressionStatement(path, state): void {
         if (
           !t.isAssignmentExpression(path.node.expression) ||
